update delete logic

+++ src/app/modules/Brand/brand.service.ts
@@ -2,7 +2,13 @@ import { Request } from 'express'
 import { Brand } from './brand.model'
 import QueryBuilder from '../../builder/QueryBuilder'
 import { searchableFields } from '../Category/category.constant'
-
+import { Product } from '../Product/product.model'
+import { AppError } from '../../Error/AppError'
+import httpStatus from 'http-status'
+import { PromotionalBanner } from '../PromotionalBanner/promotionalBanner.model'
+import { SideBanner } from '../SideBanner/sidebanner.model'
+import { Story } from '../Story/story.model'
+import { BrandOffer } from '../TopBrandsAndOffers/topBrandAndOffers.model'
 const createBrand = async (req: Request) => {
   const payload = req.body
   const file = req.file
@@ -96,6 +102,61 @@ const updateBrand = async (req: Request) => {
 
 const deleteBrand = async (id: string) => {
   const _id = id
+  const productsCount = await Product.countDocuments({
+    brand: _id,
+    isDeleted: false,
+  })
+  if (productsCount > 0) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      'Cannot delete Brand. Products are associated with this brand.'
+    )
+  }
+
+  const promotionalBannerCount = await PromotionalBanner.countDocuments({
+    brand: _id,
+    isDeleted: false,
+  })
+  if (promotionalBannerCount > 0) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      'Cannot delete Brand. Promotional Banners are associated with this brand.'
+    )
+  }
+
+  const sideBanner = await SideBanner.countDocuments({
+    brand: _id,
+    isDeleted: false,
+  })
+  if (sideBanner > 0) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      'Cannot delete Brand. Side Banners are associated with this brand.'
+    )
+  }
+
+  const storyCount = await Story.countDocuments({
+    brand: _id,
+    isDeleted: false,
+  })
+  if (storyCount > 0) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      'Cannot delete Brand. Stories are associated with this brand.'
+    )
+  }
+
+  const brandOfferCount = await BrandOffer.countDocuments({
+    brand: _id,
+    isDeleted: false,
+  })
+  if (brandOfferCount > 0) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      'Cannot delete Brand. Brand Offers are associated with this brand.'
+    )
+  }
+
   const result = await Brand.findByIdAndDelete(_id)
   return result
 }
